requirements.txt
django==4.0
psycopg2-binary==2.8.5
pillow==7.2.0
djangorestframework
markdown
django-filter